refactor -> Controller 클래스 생성
